requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
openai-whisper>=20231117
sounddevice>=0.4.6
scipy>=1.11.0
numpy>=1.24.0
